requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

babel==2.6.0              # via delorean
bottle==0.12.16
cachetools==3.0.0         # via google-auth
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
delorean==1.0.0
google-api-core[grpc]==1.7.0  # via google-cloud-core, google-cloud-datastore
google-api-python-client==1.7.7
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.2        # via google-api-core, google-api-python-client, google-auth-httplib2
google-cloud-core==0.29.1  # via google-cloud-datastore
google-cloud-datastore==1.7.3
googleapis-common-protos==1.5.5  # via google-api-core
grpcio==1.17.1            # via google-api-core
httplib2==0.12.0          # via google-api-python-client, google-auth-httplib2
humanize==0.5.1           # via delorean
idna==2.8                 # via requests
oauthlib==2.1.0           # via requests-oauthlib
protobuf==3.6.1           # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.3     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
python-dateutil==2.7.5    # via delorean
pytz==2018.7              # via babel, delorean, google-api-core, tzlocal
requests-oauthlib==1.0.0
requests==2.21.0          # via google-api-core, requests-oauthlib
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, grpcio, protobuf, python-dateutil
tzlocal==1.5.1            # via delorean
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.24.1           # via requests
